// frontend/src/app/api/generated/model/currentUserResponse.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRole } from './userRole';


export interface CurrentUserResponse { 
    id: string;
    username: string;
    role: UserRole;
}